@@ -3,11 +3,14 @@ import { Platform } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
 
-import { HomePage } from '../pages/home/home';
+import { TabsPage } from '../pages/tabs/tabs';
+import {HomePage} from "../pages/home/home";
+
 @Component({
   templateUrl: 'app.html'
 })
 export class MyApp {
+  // rootPage:any = TabsPage;
   rootPage:any = HomePage;
 
   constructor(platform: Platform, statusBar: StatusBar, splashScreen: SplashScreen) {
@@ -19,4 +22,3 @@ export class MyApp {
     });
   }
 }
-
