@@ -1,18 +1,36 @@
-import { Component } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import { NavController } from 'ionic-angular';
-import { VlcService } from "../../app/VLC"
+import {PlaylistReference, StatusReference, VlcProvider} from "../../providers/vlc/vlc";
 import {PlaylistPage} from "../playlist/playlist";
-import {HelpPage} from "../help/help";
+import {ControlPage} from "../control/control";
 
 @Component({
   selector: 'page-home',
   templateUrl: 'home.html'
 })
-export class HomePage {
+export class HomePage implements OnInit{
 
-  constructor(public navCtrl: NavController, private proxy:VlcService) {
+  public status: StatusReference;
+  public playlist: PlaylistReference;
+
+
+  constructor(public navCtrl: NavController,
+              public vlc: VlcProvider) {
+
+    this.status = vlc.status;
+    this.playlist = vlc.playlist;
   }
 
+  ngOnInit(): void {
+
+    this.vlc.getStatus();
+  }
+
+  controlOpen() {
+
+    console.log( [this], "controlOpen" );
+    ControlPage.pushOnTo( this.navCtrl );
+  }
 
   playlistOpen() {
 
@@ -22,14 +40,9 @@ export class HomePage {
 
   helpOpen() {
 
-    this.navCtrl.push( HelpPage );
+    console.log( [this], "helpOpen" );
+    this.navCtrl.push( "HelpPage" );
   }
 
-  async ngOnInit() {
-
-    let status = await this.proxy.status();
-    console.log( [this], "ngOnInit", status );
-    // this.navCtrl.push(PlaylistPage);
-  }
 
 }
